+++ apps/lib/cmp_mock_srv.c
